client front done

+++ database/migrations/2022_09_05_173118_create_claim_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('claim', function (Blueprint $table) {
             $table->id();
             $table->string('type');
-            $table->integer('client');
+            $table->bigInteger('client');
             $table->string('agent')->nullable();
             $table->string('callStatus')->nullable();
             $table->dateTime('inQueue')->nullable();
